pyproject.toml
[tool.poetry]
name = "langchain-kr"
version = "0.1.1"
description = "Ultimate LangChain Tutorial written in Korean by TeddyNote."
authors = ["teddy <teddylee777@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = ">=0.115.8,<0.116.0"
uvicorn = ">=0.34.0,<0.35.0"
python-dotenv = ">=1.0.1,<2.0.0"
langchain-community = ">=0.3.16,<0.4.0"
langchain-core = ">=0.3.33,<0.4.0"
langchain-openai = ">=0.3.3,<0.4.0"
faiss-cpu = ">=1.10.0,<2.0.0"
pydantic = ">=2.10.6,<3.0.0"
jinja2 = ">=3.1.5,<4.0.0"
langchain = ">=0.3.17,<0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

torch = { url = "https://download.pytorch.org/whl/cu121/torch-2.5.1%2Bcu121-cp311-cp311-win_amd64.whl" }
torchvision = { url = "https://download.pytorch.org/whl/cu121/torchvision-0.15.2%2Bcu121-cp311-cp311-win_amd64.whl" }
torchaudio = { url = "https://download.pytorch.org/whl/cu121/torchaudio-2.5.1%2Bcu121-cp311-cp311-win_amd64.whl" }
